refact : 테스트코드가 작동하도록 수정
