+++ src/main/java/nextstep/subway/path/dto/LinePathRequest.java
@@ -1,0 +1,15 @@
+package nextstep.subway.path.dto;
+
+
+public class LinePathRequest {
+    private Long source;
+    private Long target;
+
+    public LinePathRequest() {
+    }
+
+    public LinePathRequest(Long source, Long target) {
+        this.source = source;
+        this.target = target;
+    }
+}
