+++ src/main/java/nextstep/subway/path/dto/LinePathResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.line.dto;
+package nextstep.subway.path.dto;
 
 import nextstep.subway.station.dto.StationResponse;
